Improved types
These changes are based on the Yazi types file that I maintain together with my augment-command.yazi plugin.

The Color type now shows all the supported named colours like “blue” and “red”.

The Url global is now a function that returns a Url object, as it is not a Url object by itself and is never used that way. This allows the normal usage of `Url(“path/to/file”)` to be properly typed.

The Command global is now either the Command object or a function that returns a Command object. This allows the usage of `Command(“shell_command”)` to be properly typed.