Update installation.md with a line about mine type
I might be wrong. I found that `vnd.` was removed from the mime-type in Yazi. I originally thought yazi used the one from the `file` command, so I put the mime type from `file`, and my plugin was not working. It took me a while to figure this out, so I wanted to give a heads-up here. But I might be wrong, or this is not necessary, so feel free to ignore this. Thanks.